refactor: remove mTLS certificate from CI workflow environment
