+++ crates/forge_ci/src/lib.rs
@@ -1,3 +1,4 @@
 mod jobs;
 mod matrix;
 pub mod workflow;
+pub mod workflows;
